+++ docs/www.fpart.org/docs/changelog.md
@@ -13,6 +13,9 @@
       special partition '0' has been used
     - fpart: avoid returning special partition '0' when it is empty
       (partly fixes GH issue #36)
+    - fpart: add option -S to skip, in live mode, files that are bigger than
+      maximum partition size
+      (see man page for more details, fixes GH issue #36)
     - fpsync: add more checks for incompatible rsync options
     - fpsync: force non-recursive rsync(1) in file-based mode
 ```